(feat) Upgrade to use now.sh and major refactor
