--- .travis.yml
@@ -1,14 +1,0 @@
-language: node_js
-node_js:
-  - "9.3.0"
-  
-script:
-  - gulp lambda:package
-
-cache:
-  directories:
-  - node_modules
-
-notifications:
-  email:
-    - jeremyrajan@gmail.com
