--- debug.js
@@ -1,3 +1,0 @@
-const lambdaTypescript = require('aws-lambda-typescript');
-
-lambdaTypescript.runLocalServer(__dirname);
