--- gulpfile.js
@@ -1,4 +1,0 @@
-const gulp = require('gulp');
-const awsLambdaTypescript = require('aws-lambda-typescript');
-
-awsLambdaTypescript.registerBuildGulpTasks(gulp, __dirname);
