+++ now.json
@@ -1,0 +1,6 @@
+{
+  "routes": [
+    { "src": "/api", "dest": "/api/index.ts" },
+    { "src": "/api/badge.svg", "dest": "/api/index.ts" }
+  ]
+}
